+++ src/Domain/Tests/Fixtures/TestOtherDomainUuid.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace MsgPhp\Domain\Tests\Fixtures;
 
 use MsgPhp\Domain\DomainId;
-use MsgPhp\Domain\Infrastructure\Uuid\DomainIdTrait;
+use MsgPhp\Domain\Infrastructure\Uid\DomainIdTrait;
 
 final class TestOtherDomainUuid implements DomainId
 {
